refactor(firmware): make the plant model single precision [SOL-PLANT-STRUCTURE-SEAM-FIRST-STRUCTURE.C1, .C2, .C3, .C4]
Double precision was chosen without a reason and quietly narrowed the field of
controllers the machine could be built on. Single precision is handled natively
by every part that would reasonably be chosen; double either needs a part with a
double-precision unit or is emulated in software on one without. Leaving the
part choice open is worth more here than a precision nothing asks for.

Nothing wanted the extra width. The measurement is a twelve- to fifteen-bit
conversion, and the control-design obligation expects a first identification to
recover the response shape to within tens of percent. Single precision carries
about seven significant figures against a model whose error is O(10%). Running
the same nominal description through both builds agrees to seven figures --
38.081510 against 38.081512 -- which is five orders of magnitude finer than the
model's own fidelity.

Compiled for the nominated Cortex-M4F, the structure emits 30 hardware f32
instructions and no soft-float helper call, against 30 helper calls before;
the parameter loader likewise, now that it reads with strtof rather than
strtod. Both are static properties of the emitted code rather than timing
claims: the real-time obligation requires those be measured on the target that
is eventually selected.

A value the type cannot hold -- too large, or so small it would arrive as zero
-- is now refused rather than delivered as infinity or as zero.

The coefficient-sensitivity test was re-checked against four hard-coding
mutations at the narrower width and still catches all four, so the reduced
precision has not blinded it.

+++ firmware/include/plant_model.h
@@ -77,6 +77,6 @@ bool plant_model_step(plant_model_t *model, const plant_actuation_t *actuation,
  * null, or when the quantity is not one of the enumerated ones.
  */
 bool plant_model_quantity(const plant_model_t *model, plant_quantity_t quantity,
-                          double *value);
+                          float *value);
 
 #endif /* PLANT_MODEL_H */
